pyproject.toml
[tool.poetry]
name = "mypynet"
version = "0.1.0"
description = ""
authors = ["Palltar <begunkovich.pavel@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
djangorestframework = "^3.12.4"
Pillow = "^8.2.0"
psycopg2 = "^2.9.1"
djoser = "^2.1.0"
djangorestframework-simplejwt = "^4.7.1"
drf-yasg = "^1.20.0"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
